feat: centralize date formatting
